Add run full and major changes

+++ source/cost-functions.h
@@ -2,15 +2,15 @@
 
 #include "graph.h"
 
-int dijkstra(int searchNodeCost, Node current, Node start, Node end) {
+double dijkstra(double searchNodeCost, Node current, Node start, Node end) {
 	return searchNodeCost;
 }
 
-int astar(int searchNodeCost, Node current, Node start, Node end) {
+double astar(double searchNodeCost, Node current, Node start, Node end) {
 	return searchNodeCost + current.realDistance(end);
 }
 
-int greedy(int searchNodeCost, Node current, Node start, Node end) {
+double greedy(double searchNodeCost, Node current, Node start, Node end) {
 	return current.realDistance(end);
 }
